requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlalchemy>=2.0.0  # Para o indexador
pytest-asyncio>=0.21.0  # Para testes assíncronos
feedparser>=6.0.0
pyyaml>=6.0.0
newspaper3k
langchain-openai
tenacity>=8.2.3